+++ packages/hydrogen/CHANGELOG.md
@@ -10,6 +10,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - fix: default to `retry: false` in `useQuery`
 - fix: starter template media gallery error when handling videos
 - fix: add 404 link to footer
+- fix: align font styles for h1 and paragraph
 
 ## 0.6.4 - 2021-11-11